Impl save file upgrading. Fix bug on cycle removal
Old save files (API version 1) will be detected and upgraded to the new model format.

When removing a cycle, if sorties were slaved to that cycle, this
would cause an error. FIxed by resetting all sorties slaved to
a cycle to not reference a cycle when that cycle is deleted.